Refactor(#151): classInfo 받아오도록 수정
