+++ src/entities/Class/ClassInfo/index.tsx
@@ -4,21 +4,21 @@ import { ClassInfoProps } from '@/shared/types/Class/classroom';
 import ProgressBar from '@/entities/UI/ProgressBar';
 import { FaUserAlt } from "react-icons/fa";
 import { getClassInfo } from '../api';
+import HaeyulImg from '@/../public/Paletto/Haeyul.png';
 import * as s from './styles';
 
 export const ClassInfo: React.FC<ClassInfoProps> = ({
-  name, teacherName, description, progress, maxProgress, imageUrl
+  name, teacherName, description, progress, maxProgress
 }) => {
   const { classId } = useParams<{ classId: string }>();
   const [classData, setClassData] = useState({
     name: name || '',
-    teacherName: teacherName || '',
+    teacherName: teacherName || "유근찬",
     description: description || '',
     progress: progress || 0,
     maxProgress: maxProgress || 100,
-    imageUrl: imageUrl || ''
   });
-  const [isLoading, setIsLoading] = useState(true);
+  const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
     if (classId) {
@@ -28,15 +28,21 @@ export const ClassInfo: React.FC<ClassInfoProps> = ({
 
   const loadClassInfo = async () => {
     if (!classId) return;
-    
+
     try {
       setIsLoading(true);
-      
+
       const response = await getClassInfo(classId);
-      console.log("넘어온 값",response)
-      
+      console.log("API response:", response);
+
       if (response && typeof response === 'object') {
-        setClassData(response);
+        setClassData({
+          name: response.name || "",
+          description: response.description || "",
+          teacherName: response.teacherName || "",
+          progress: 0,
+          maxProgress: 100,
+        });
       } else {
         console.warn('클래스 정보 조회 실패:', response);
       }
@@ -51,7 +57,7 @@ export const ClassInfo: React.FC<ClassInfoProps> = ({
     return (
       <s.Container>
         {/* 나중에 loading style로 변경 */}
-        <div> 
+        <div>
           클래스 정보를 불러오는 중...
         </div>
       </s.Container>
@@ -66,12 +72,12 @@ export const ClassInfo: React.FC<ClassInfoProps> = ({
 
         <s.TeacherRow>
           <FaUserAlt />
-          <span>{classData.teacherName}</span>
+          <span>{classData.teacherName}님</span>
         </s.TeacherRow>
 
         <ProgressBar progress={classData.progress} maxProgress={classData.maxProgress} />
       </s.LeftSection>
-      {classData.imageUrl && <s.Img imageUrl={classData.imageUrl} />}
+      <s.Img imageUrl={HaeyulImg} />
     </s.Container>
   );
 };
